requirements.txt
grpcio>=1.12.0
protobuf>=3.6.0
yagrc>=1.1.1
